Add autoCompletion for better abstraction.
